move some structs to new traversal.rs
